Moved classes into packages
