+++ src/main/java/com/twistylime/bossLand/metrics/Metrics.java
@@ -25,7 +25,7 @@
  * authors and contributors and should not be interpreted as representing official policies,
  * either expressed or implied, of anybody else.
  */
-package com.twistylime.bossLand;
+package com.twistylime.bossLand.metrics;
 
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.InvalidConfigurationException;
